+++ tomcat/src/main/java/org/apache/coyote/servlet/servlets/AbstractServlet.java
@@ -2,6 +2,7 @@ package org.apache.coyote.servlet.servlets;
 
 import org.apache.catalina.SessionManager;
 import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.request.header.Method;
 import org.apache.coyote.http11.response.HttpResponse;
 
 public abstract class AbstractServlet implements Servlet {
@@ -13,7 +14,23 @@ public abstract class AbstractServlet implements Servlet {
     }
 
     @Override
-    public abstract void service(final HttpRequest httpRequest, final HttpResponse httpResponse);
+    public void service(final HttpRequest httpRequest, final HttpResponse httpResponse) {
+        final Method method = httpRequest.getMethod();
+
+        if (method.isGet()) {
+            doGet(httpRequest, httpResponse);
+            return;
+        }
+        if (method.isPost()) {
+            doPost(httpRequest, httpResponse);
+            return;
+        }
+        setNotFound(httpResponse);
+    }
+
+    protected abstract void doGet(final HttpRequest httpRequest, final HttpResponse httpResponse);
+
+    protected abstract void doPost(final HttpRequest httpRequest, final HttpResponse httpResponse);
 
     protected void setUnauthorized(final HttpResponse httpResponse) {
         httpResponse.setStatusCode("401")
