+++ tomcat/src/main/java/org/apache/coyote/servlet/servlets/LoginServlet.java
@@ -5,7 +5,6 @@ import java.util.Optional;
 
 import org.apache.catalina.SessionManager;
 import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.request.header.Method;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -22,21 +21,7 @@ public class LoginServlet extends AbstractServlet {
     }
 
     @Override
-    public void service(final HttpRequest httpRequest, final HttpResponse httpResponse) {
-        final Method method = httpRequest.getMethod();
-
-        if (method.isGet()) {
-            doGet(httpRequest, httpResponse);
-            return;
-        }
-        if (method.isPost()) {
-            doPost(httpRequest, httpResponse);
-            return;
-        }
-        setNotFound(httpResponse);
-    }
-
-    private void doGet(final HttpRequest httpRequest, final HttpResponse httpResponse) {
+    protected void doGet(final HttpRequest httpRequest, final HttpResponse httpResponse) {
         if (sessionManager.isLoginAccount(httpRequest)) {
             httpResponse.setStatusCode("302")
                 .setLocation("/index.html");
@@ -46,7 +31,8 @@ public class LoginServlet extends AbstractServlet {
             .setBody("/login.html");
     }
 
-    private void doPost(final HttpRequest httpRequest, final HttpResponse httpResponse) {
+    @Override
+    protected void doPost(final HttpRequest httpRequest, final HttpResponse httpResponse) {
         final Map<String, String> bodies = httpRequest.getBodies();
         if (isNotContainEssentialParams(bodies)) {
             setNotFound(httpResponse);
