+++ tomcat/src/main/java/org/apache/coyote/servlet/servlets/RegisterServlet.java
@@ -4,7 +4,6 @@ import java.util.Map;
 
 import org.apache.catalina.SessionManager;
 import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.request.header.Method;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -20,26 +19,13 @@ public class RegisterServlet extends AbstractServlet {
     }
 
     @Override
-    public void service(final HttpRequest httpRequest, final HttpResponse httpResponse) {
-        final Method method = httpRequest.getMethod();
-
-        if (method.isGet()) {
-            doGet(httpRequest, httpResponse);
-            return;
-        }
-        if (method.isPost()) {
-            doPost(httpRequest, httpResponse);
-            return;
-        }
-        setNotFound(httpResponse);
-    }
-
-    private void doGet(final HttpRequest httpRequest, final HttpResponse httpResponse) {
+    protected void doGet(final HttpRequest httpRequest, final HttpResponse httpResponse) {
         httpResponse.setStatusCode("200")
             .setBody("/register.html");
     }
 
-    private void doPost(final HttpRequest httpRequest, final HttpResponse httpResponse) {
+    @Override
+    protected void doPost(final HttpRequest httpRequest, final HttpResponse httpResponse) {
         final Map<String, String> bodies = httpRequest.getBodies();
         UserService.save(bodies.get("account"), bodies.get("password"), bodies.get("email"));
         log.info("register success to ID : {}", bodies.get("account"));
