+++ tomcat/src/main/java/org/apache/coyote/servlet/servlets/ResourceServlet.java
@@ -2,7 +2,6 @@ package org.apache.coyote.servlet.servlets;
 
 import org.apache.catalina.SessionManager;
 import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.request.header.Method;
 import org.apache.coyote.http11.response.HttpResponse;
 
 public class ResourceServlet extends AbstractServlet {
@@ -12,18 +11,13 @@ public class ResourceServlet extends AbstractServlet {
     }
 
     @Override
-    public void service(final HttpRequest httpRequest, final HttpResponse httpResponse) {
-        final Method method = httpRequest.getMethod();
-
-        if (method.isGet()) {
-            doGet(httpRequest, httpResponse);
-            return;
-        }
-        setNotFound(httpResponse);
-    }
-
-    private void doGet(final HttpRequest httpRequest, final HttpResponse httpResponse) {
+    protected void doGet(final HttpRequest httpRequest, final HttpResponse httpResponse) {
         httpResponse.setStatusCode("200")
             .setBody(httpRequest.getUrl());
+    }
+
+    @Override
+    protected void doPost(final HttpRequest httpRequest, final HttpResponse httpResponse) {
+        setNotFound(httpResponse);
     }
 }
